+++ src/test/java/study/step3/RaceRecordTest.java
@@ -6,12 +6,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import java.util.AbstractMap.SimpleEntry;
-import java.util.*;
 import java.util.concurrent.atomic.AtomicInteger;
-import java.util.function.BiConsumer;
 
-import static java.util.stream.Collectors.toSet;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static study.step3.CarRacingTest.TestingCar;
@@ -102,66 +98,6 @@ public class RaceRecordTest {
         for (Car car : cars) {
             car.move();
             raceRecord.saveRecord(car);
-        }
-    }
-
-    private static class RaceRecord {
-        private final Map<String, List<Boolean>> records = new TreeMap<>();
-
-        public void saveRecord(Car car) {
-            records.computeIfAbsent(car.getName(), key -> new ArrayList<>()) //
-                    .add(car.isMoved());
-        }
-
-        public Set<String> listMostMovingNames() {
-            int mostMove = getMostMove();
-            return records.entrySet() //
-                    .stream() //
-                    .map(entry -> new SimpleEntry<>(entry.getKey(), countMove(entry.getValue()))) //
-                    .filter(entry -> entry.getValue() >= mostMove) //
-                    .map(Map.Entry::getKey) //
-                    .collect(toSet());
-        }
-
-        private int getMostMove() {
-            return records.values() //
-                    .stream() //
-                    .mapToInt(this::countMove) //
-                    .max() //
-                    .orElse(0);
-        }
-
-        private int countMove(List<Boolean> values) {
-            return values.stream().mapToInt(moved -> moved ? 1 : 0).sum();
-        }
-
-        public int getTotalTry() {
-            if (records.isEmpty()) {
-                return 0;
-            }
-
-            int totalTry = records.values().iterator().next().size();
-
-            checkAllRecord(totalTry);
-
-            return totalTry;
-        }
-
-        public void forEachRecordUntil(int tries, BiConsumer<String, List<Boolean>> biConsumer) {
-            records.forEach((name, record) -> biConsumer.accept(name, record.subList(0, tries)));
-        }
-
-        private void checkAllRecord(int expectedTry) {
-            records.values() //
-                    .stream() //
-                    .map(List::size) //
-                    .forEach(tries -> throwIfNotEqual(expectedTry, tries));
-        }
-
-        private void throwIfNotEqual(int expectedTry, int tries) {
-            if (expectedTry != tries) {
-                throw new IllegalStateException("모든 자동차의 이동시도횟수 기록은 동일해야 합니다.");
-            }
         }
     }
